requirements.txt
pyyaml>=6.0
rich>=13.0
jinja2>=3.1
invoke>=2.0
